added roboinfo page

+++ src/components/hooks/useHire.tsx
@@ -1,21 +1,26 @@
 import { robotType } from '../RobotList/RobotList';
-import { useRobo } from './useRobo';
 
-export const useHire = async (robotId: string) => {
-    const addCart = async (newRobot: robotType) => {
-        await fetch('http://localhost:3000/cart', {
+export const useHire = () => {
+    const addCart = async (newRobot: any) => {
+        const response = await fetch('http://localhost:3000/robots');
+        const robots = await response.json();
+        const targetRobot = robots.find(
+            (robot: any) => newRobot.id === robot.id
+        );
+
+        if (targetRobot) {
+            await fetch(`http://localhost:3000/robots/${newRobot.id}`, {
+                method: 'DELETE',
+            });
+        }
+
+        await fetch('http://localhost:3000/hired', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
             },
             body: JSON.stringify(newRobot),
         });
-        const response = await fetch(`http://localhost:3000/robots/${robotId}`);
-        const robot = await response.json();
-        const updatedRobot = { ...robot, hired: true };
-        await fetch(`http://localhost:3000/robots/${robotId}`, {
-            method: 'DELETE',
-        });
-        addCart(updatedRobot);
     };
+    return { addCart };
 };
